Add PyTorchObjectDetector
